refactor: move registry to its own source file
