+++ src/execution/mod.rs
@@ -19,6 +19,7 @@ pub mod hooks;
 mod interpreter_loop;
 pub(crate) mod linear_memory;
 pub(crate) mod locals;
+pub mod registry;
 pub mod store;
 pub mod value;
 pub mod value_stack;
